# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/providers": {
            "get": {
                "description": "分页获取所有邮箱提供商配置列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider管理"
                ],
                "summary": "分页获取 Provider 列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/fusionmail_internal_model.ProviderResponse"
                                                    }
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "page_size": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "format": "int64"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "为系统添加新的邮箱提供商配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider管理"
                ],
                "summary": "创建 Provider 配置",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_model.Provider"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/providers/all": {
            "get": {
                "description": "获取所有邮箱提供商配置列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider管理"
                ],
                "summary": "获取 Provider 列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/providers/{id}": {
            "get": {
                "description": "通过ID获取指定的邮箱提供商配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider管理"
                ],
                "summary": "获取 Provider 详情",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Provider ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定ID的邮箱提供商配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider管理"
                ],
                "summary": "更新 Provider 配置",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Provider ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_model.Provider"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定ID的邮箱提供商配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider管理"
                ],
                "summary": "删除 Provider 配置",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Provider ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fusionmail_internal_model.Provider": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "时间戳",
                    "type": "string"
                },
                "description": {
                    "description": "描述信息",
                    "type": "string"
                },
                "display_name": {
                    "description": "显示名称",
                    "type": "string"
                },
                "enabled": {
                    "description": "管理字段",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "imap_host": {
                    "description": "服务器配置",
                    "type": "string"
                },
                "imap_port": {
                    "description": "IMAP端口",
                    "type": "integer"
                },
                "metadata": {
                    "description": "JSON格式的额外配置",
                    "type": "string"
                },
                "name": {
                    "description": "基础信息",
                    "type": "string"
                },
                "pop3_host": {
                    "description": "POP3服务器地址",
                    "type": "string"
                },
                "pop3_port": {
                    "description": "POP3端口",
                    "type": "integer"
                },
                "provider_type": {
                    "description": "提供商类型（枚举值）",
                    "type": "integer"
                },
                "recommended_protocol": {
                    "description": "推荐协议",
                    "type": "string"
                },
                "requires_oauth": {
                    "description": "是否强制OAuth",
                    "type": "boolean"
                },
                "smmtp_port": {
                    "description": "SMTP端口（预留）",
                    "type": "integer"
                },
                "smtp_host": {
                    "description": "SMTP服务器地址（预留）",
                    "type": "string"
                },
                "sort_order": {
                    "description": "排序顺序",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "fusionmail_internal_model.ProviderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "imap_host": {
                    "type": "string"
                },
                "imap_port": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pop3_host": {
                    "type": "string"
                },
                "pop3_port": {
                    "type": "integer"
                },
                "provider_type": {
                    "description": "提供商类型",
                    "type": "integer"
                },
                "recommended_protocol": {
                    "type": "string"
                },
                "requires_oauth": {
                    "type": "boolean"
                },
                "smtp_host": {
                    "type": "string"
                },
                "smtp_port": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                },
                "supported_protocols": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}